Compensate for CustomCommandExecutor behavior:
CustomCommandExecutor allows us to ASK for whatever parameters we want.
However, when we ask for an array or varargs, then CustomCommandExecutor
is NOT giving us the remaining command-line arguments. But instead, it is
giving us ALL the command-line arguments.

From reading thru the code, this is in fact intentional, eventho
un-expected. And we therefore cannot change this behavior without breaking
all code already reliant on the current behavior.

Instead, we have to compensate for this by parsing the String[] args
ourselves.